Decrease bundle size by removing a lot of deps
