+++ .github/workflows/publish.yaml
@@ -32,6 +32,9 @@ jobs:
       - name: Publish script package
         run: cd scripts/ && npx jsr publish
 
+      - name: Publish dev package
+        run: cd dev/ && npx jsr publish
+
       - name: Set up QEMU
         uses: docker/setup-qemu-action@v3
