+++ .github/workflows/releaser.yaml
@@ -152,10 +152,12 @@ jobs:
         run: |
           jq --arg new_version "${{ steps.determine_version.outputs.new_version }}" '.version = $new_version' deno.json > tmp.$$.json && mv tmp.$$.json deno.json
           jq --arg new_version "${{ steps.determine_version.outputs.new_version }}" '.version = $new_version' ./scripts/deno.json > tmp.$$.json && mv tmp.$$.json ./scripts/deno.json
+          jq --arg new_version "${{ steps.determine_version.outputs.new_version }}" '.version = $new_version' ./dev/deno.json > tmp.$$.json && mv tmp.$$.json ./dev/deno.json
+
 
           git config user.name "decobot"
           git config user.email "github-actions[bot]@users.noreply.github.com"
-          git add deno.json ./scripts/deno.json
+          git add deno.json ./scripts/deno.json ./dev/deno.json
           git commit -m "Update version to ${{ steps.determine_version.outputs.new_version }}"
           git push origin main
