+++ scripts/deno.json
@@ -19,7 +19,7 @@
   },
   "imports": {
     "@cliffy/prompt": "jsr:@cliffy/prompt@^1.0.0-rc.5",
-    "@deco/codemod-toolkit": "jsr:@deco/codemod-toolkit@^0.3.1",
+    "@deco/codemod-toolkit": "jsr:@deco/codemod-toolkit@^0.3.3",
     "@std/flags": "jsr:@std/flags@^0.224.0",
     "@std/fmt": "jsr:@std/fmt@^1.0.0",
     "@std/fs": "jsr:@std/fs@^1.0.1",
